refactor(webview): extract ScanPublishingHandler from webview-provider
Moves handleStartScan plus the URL-canonicalization, endpoint-merging,
local-finding-aggregation, and aggressive-suggestion helpers it depends
on (~470 lines) into src/webview/scan-publishing-handler.ts. The handler
calls back into the provider through a ScanPublishingHandlerContext for
state mutations, key validation updates, project resolution, and the
debug-export side effect.

Structurally pure extraction — no behavioral change. webview-provider.ts
drops from 1534 to 791 lines (under the ~800 target set by the plan).
